requirements.txt
opencv-python
numpy
torch
torchvision
Pillow
wandb
tqdm
mkl-service
scikit-learn
seaborn
matplotlib
